Fix remaining build errors in kernel subdir
A lot of this is sloppy, and there are still a ton of warnings, but
everything in kernel/ produces a .o now.

There's a ton of asm code that I just completely commented out for now.
I also left kprtinf in a nonsensical state just to get it to build, and
I tweaked _sbrk a bit.  I think the latter is OK, but could be important
to note that I changed it if I run into trouble there later.  There
might be other stuff I clobbered, but that's all I'm seeing from a quick
glance at the diff...