Link all built objects
I realized that a bunch of my linker errors were because I just wasn't
linking all of the objects, which seems stupid in retrospect.  Anyway,
this links everything that is built, and tweaks the linker script to
avoid errors related to the fact that the CODE section gets way too
large when we build everything.

This will need to be fixed up somehow with some more clever bank usage.
The modified linker script at least doesn't generate errors, but I don't
think it will produce a valid NES ROM image.